chore: refactors  domain respository layer
