+++ backend/src/domain/repository/comment.type.ts
@@ -1,4 +1,4 @@
 import { DomainComment } from '../entities/comment.model';
-import { IGenericDomainRepository } from './domainRepository.interface';
+import { IGenericDomainRepository } from '../utils/interfaces/domainRepository.interface';
 
 export type CommentRepository = IGenericDomainRepository<DomainComment>;
