+++ backend/src/domain/repository/profile.type.ts
@@ -1,4 +1,4 @@
 import { DomainProfile } from '../entities/profile.model';
-import { IGenericDomainRepository } from './domainRepository.interface';
+import { IGenericDomainRepository } from '../utils/interfaces/domainRepository.interface';
 
 export type ProfileRepository = IGenericDomainRepository<DomainProfile>;
