+++ backend/src/domain/repository/topic.type.ts
@@ -1,4 +1,4 @@
 import { DomainTopic } from '../entities/topic.model';
-import { IGenericDomainRepository } from './domainRepository.interface';
+import { IGenericDomainRepository } from '../utils/interfaces/domainRepository.interface';
 
 export type TopicRepository = IGenericDomainRepository<DomainTopic>;
