+++ backend/src/domain/repository/user.type.ts
@@ -1,4 +1,4 @@
 import { DomainUser } from '../entities/user.model';
-import { IGenericDomainRepository } from './domainRepository.interface';
+import { IGenericDomainRepository } from '../utils/interfaces/domainRepository.interface';
 
 export type UserRepository = IGenericDomainRepository<DomainUser>;
